requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
cryptography>=42
qrcode[pil]>=7.4
PyJWT>=2.8
email-validator>=2.1
psycopg[binary]>=3.2
